# src/backend/api/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GetGPX API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "This GetGPX deployment"
    }
  ],
...
  "paths": {
    "/tracks": {
      "post": {
        "operationId": "upload",
        "summary": "Upload a track",
        "description": "Requires a website session. Source is stored durably before 202; processing continues asynchronously. Poll the Location URL. User tier quota applies (BASIC: 100, PREMIUM: 1,000 by default).",
        "tags": [
          "Track management"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Relative URL of the owner-only processing status endpoint.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid website session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured limit (GPX: 25 MiB; JSON: 16 KiB).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported GPX content type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "User track quota reached (TRACK_LIMIT_REACHED).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-GPX-Filename",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Percent-encoded filename ending in .gpx, up to 255 decoded characters. Path components are removed."
          },
          {
            "in": "header",
            "name": "X-Track-Type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RouteType"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Raw GPX bytes, not multipart/form-data. Maximum 25 MiB and 500,000 accepted route points.",
          "content": {
            "application/gpx+xml": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "application/xml": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "removeMany",
        "summary": "Delete selected tracks",
        "description": "Permanently deletes owned tracks and stored objects. Missing or non-owned IDs are omitted from deletedIds. Duplicate IDs are processed once.",
        "tags": [
          "Track management"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedIdsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON or invalid cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid website session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured limit (GPX: 25 MiB; JSON: 16 KiB).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported JSON charset or encoding.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ids"
              }
            }
          }
        }
      }
    },
...
    "/tracks/{id}": {
      "get": {
        "operationId": "publicTrack",
        "summary": "Read track metadata",
        "description": "Lightweight point-free metadata and complete precomputed summary from MongoDB. No S3 read. Geometry is only available from /analysis; there is no include=geometry option. A processing or failed track still returns 200 with its available information. Missing measurements are null. summary may be null before analysis is available.",
        "tags": [
          "Tracks"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Track not found or not owned by the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PublicId"
            }
          }
        ]
      },
      "patch": {
        "operationId": "update",
        "summary": "Edit track metadata",
        "description": "Owner only. Compatibility note: despite PATCH, title, routeType and speedKmh must all be supplied. Omitting externalLinks clears links. Changes speed/duration without re-analysis. Updating an initial processing track is allowed; replacing an active track temporarily prevents editing.",
        "tags": [
          "Track management"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedTrackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON or invalid cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid website session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Track not found or not owned by the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Processing conflict or analysis not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured limit (GPX: 25 MiB; JSON: 16 KiB).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported JSON charset or encoding.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PublicId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrack"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete a track",
        "description": "Owner only. Permanently deletes metadata and stored objects. Repeating deletion of a missing ID succeeds. A track owned by someone else returns 404.",
        "tags": [
          "Track management"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid website session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Track not found or not owned by the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PublicId"
            }
          }
        ]
      }
    },
...
    "/tracks/{id}/gpx": {
      "put": {
        "operationId": "replace",
        "summary": "Replace the original GPX",
        "description": "Owner only. Raw file replacement keeps the public ID and metadata. Existing active source and analysis remain available until the new revision publishes atomically. 409 also covers a missing or non-owned track with a syntactically valid ID.",
        "tags": [
          "Track management"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid website session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Track not found or not owned by the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Processing conflict or analysis not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured limit (GPX: 25 MiB; JSON: 16 KiB).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PublicId"
            }
          },
          {
            "in": "header",
            "name": "X-GPX-Filename",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Percent-encoded filename ending in .gpx, up to 255 decoded characters. Path components are removed."
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Raw GPX bytes, not multipart/form-data. Maximum 25 MiB and 500,000 accepted route points.",
          "content": {
            "application/gpx+xml": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "application/xml": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "download",
        "summary": "Download the original GPX",
        "description": "Original bytes with an attachment filename. Active source, or the initial source before first publication. The frontend/API URLs stay on the same origin even when the deployment proxies delivery through CloudFront. HEAD is also supported.",
        "tags": [
          "Tracks"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "GPX file.",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "Attachment with sanitized filename and UTF-8 filename*."
              }
            },
            "content": {
              "application/gpx+xml": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Track not found or not owned by the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Processing conflict or analysis not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Stored file could not be read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "File delivery unavailable; check reverse-proxy configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PublicId"
            }
          }
        ]
      }
    },
    "/tracks/{id}/analysis": {
      "get": {
        "operationId": "analysis",
        "summary": "Read detailed analysis",
        "description": "Detailed JSON with route points and elevation profile, streamed from S3 (or proxied by Nginx). Not wrapped in data. Active revision, with diagnostic fallback if there is no active result. Compare revision with the metadata response and refetch metadata if it changed. Storage snapshots retain calculated title/speed; apply current metadata title and metrics for presentation. HEAD is also supported.",
        "tags": [
          "Tracks"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDocument"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin browser request refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Track not found or not owned by the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Processing conflict or analysis not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or persistence failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Stored file could not be read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "File delivery unavailable; check reverse-proxy configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed for this resource.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PublicId"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "getgpx_session",
        "description": "Existing HttpOnly website session. Sign in on the site; browser requests on this origin send it automatically. No token issuance endpoint."
      }
    },
    "schemas": {
...
      "RouteType": {
        "type": "string",
        "enum": [
          "cycling",
          "road-cycling",
          "gravel-cycling",
          "mountain-biking",
          "hiking",
          "running",
          "walking",
          "driving",
          "motorcycling",
          "swimming",
          "winter-sports",
          "other"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable error code."
              },
              "message": {
                "type": "string",
                "description": "Optional human-readable legacy message. Do not parse or depend on its language."
              }
            },
            "required": [
              "code"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Ids": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicId"
            },
            "minItems": 1,
            "maxItems": 100
          }
        },
        "required": [
          "ids"
        ],
        "additionalProperties": true
      },
...
      "TrackResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Track"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
      "ProcessingResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Processing"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
      "DeletedIdsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeletedIds"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
    }
  }
}